Add stress-ng to simulate node
The pods are now distributed across the cluster.  Run an additional
stress-ng container in each pod. This does a rough simualation of
synthetic load. This is more along the lines of noisy neighbour.

Note: Currently the load run within the POD and not the container
itself. Longer term the we would like the load to run within the
container itself. This would allow the constraints of the container
that is responding to the request to be impacted by the load.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>